--- src/test/java/com/thealgorithms/strings/WordUpperCase.java
@@ -1,24 +1,0 @@
-package com.thealgorithms.strings;
-
-public class WordUpperCase {
-
-    @Test
-    public static String toUpperCase(String str) {
-        StringBuilder sb = new StringBuilder("");
-
-        char ch = Character.toUpperCase(str.charAt(0));
-        sb.append(ch);
-
-        for (int i = 1; i < str.length(); i++) {
-            if (str.charAt(i) == ' ' && i < str.length() - 1) {
-                sb.append(str.charAt(i));
-                i++;
-                sb.append(Character.toUpperCase(str.charAt(i)));
-
-            } else {
-                sb.append(str.charAt(i));
-            }
-        }
-        return sb.toString();
-    }
-}
