+++ src/test/java/com/thealgorithms/strings/WordUpperCaseTest.java
@@ -1,0 +1,20 @@
+package com.thealgorithms.strings;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+
+public class WordUpperCaseTest {
+
+    @Test
+    public void toUpperCase() {
+        String input1 = "hello world";
+        String input2 = "HelLO WoRld";
+        String input3 = "HELLO WORLD";
+
+        assertEquals("hello world", WordUpperCase.toUpperCase(input1));
+        assertEquals("hello world", WordUpperCase.toUpperCase(input2));
+        assertEquals("hello world", WordUpperCase.toUpperCase(input3));
+    }
+}
